o sdl     - implemented SdlPlay     - simplified and more efficient SDL interface     - renamed PixelTV to SdlDebug     - SdlDebug implies debugging is allowed so removed AllowDebugging 	request from gui interface     - removed stability code from SdlDebug
o television
    - added stability detection to base television implementation
    - added top/bottom scanline figures to specification types. more
	intuitive to work with in some contexts

+++ performance/fps.go
@@ -6,6 +6,6 @@ import "gopher2600/television"
 // frames-per-second and the accuracy of that value as a percentage.
 func CalcFPS(ftv television.Television, numFrames int, duration float64) (fps float64, accuracy float64) {
 	fps = float64(numFrames) / duration
-	accuracy = 100 * float64(numFrames) / (duration * ftv.GetSpec().FramesPerSecond)
+	accuracy = 100 * float64(numFrames) / (duration * float64(ftv.GetSpec().FramesPerSecond))
 	return fps, accuracy
 }
